Find a good combination of synchronization primitives
Essentially: `Arc<Mutext<Vec<_>>>`. May still be passing it around a little pointlessly, but this is a learning experience.
Passing it into `Hub` now, which means I can't derive `Default` anymore unfortunately so must fill in the ctor.

This also significantly complicates the traits for the callbacks, and all in the name of not making explicit clones of `main_state` in main.rs.
Should probably do something different, but I don't want all the meaty logic to really reside inside main.rs anyway.

Need to be able to test the logic, and for that it needs to live elsewhere.