Implement hyper::server::Handler manually
reroute implements it itself so we can't pass state through it.
Better to just have a small SeverHandle struct with our data and do simple routing ourself.

Because of this we could rewrite all the webhook handler methods to reference state on `self`.
This simplified signatures of a bunch of functions and will allow easier use of the github client later on.